refactor: change spread translatedPaths on all over the pages
