+++ web/src/infra/i18n/index.ts
@@ -4,4 +4,5 @@ export * from './hocs';
 export * from './hooks';
 export * from './utils';
 export * from './types';
+export * from './paths';
 export type { WithTranslation } from 'next-i18next';
