+++ web/src/infra/i18n/paths.ts
@@ -1,0 +1,6 @@
+import { playPaths } from 'pages/play/paths';
+import { TranslatedPath } from './types';
+
+const translatedPaths: TranslatedPath[] = [...playPaths];
+
+export default translatedPaths;
