--- web/src/infra/i18n/translatedPaths.ts
@@ -1,13 +1,0 @@
-export interface TranslatedPath {
-  original: string;
-  locales: { [locale: string]: string };
-}
-
-const translatedPaths: TranslatedPath[] = [
-  {
-    original: '/play/:rest+',
-    locales: { pt: '/jogar/:rest+' },
-  },
-];
-
-export default translatedPaths;
