+++ web/src/infra/i18n/types/index.ts
@@ -12,3 +12,8 @@ export type NextRouter = Router;
 export type { WithRouterProps } from 'next/dist/client/with-router';
 
 export type WithTranslate<T extends string = 'translate'> = Record<T, TFunction>;
+
+export interface TranslatedPath {
+  original: string;
+  locales: { [locale: string]: string };
+}
