+++ web/src/pages/play/paths.ts
@@ -1,0 +1,8 @@
+import { TranslatedPath } from 'infra/i18n/types';
+
+export const playPaths: TranslatedPath[] = [
+  {
+    original: '/play/:rest+',
+    locales: { pt: '/jogar/:rest+' },
+  },
+];
